add file-manager api
