--- packages/gatsby-theme-store/src/sdk/img/useScaledImage.ts
@@ -1,6 +1,0 @@
-import { useMemo } from 'react'
-
-import { scaleFileManagerImage } from './fileManager'
-
-export const useScaledImage = (src: string, width: number, height: number) =>
-  useMemo(() => scaleFileManagerImage(src, width, height), [height, src, width])
